backend/pyproject.toml
[project]
name = "portfolioapi"
version = "0.1.0"
description = "Backend do Portfólio Profissional Pablo Ortiz"
authors = [
    {name = "pablomigueldias", email = "pablo.miguel.dias@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "google-genai (>=1.62.0,<2.0.0)"
]

[tool.poetry]
package-mode = false 

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"